# Pertemuan 11/Restfull API/app/Entities/Observers.php
<?php namespace App\Entities;
use asligresik\easyapi\Entities\BaseEntity;
/**    
* Class Observers
* @OA\Schema(
*     title="Observers",
*     description="Observers"
* )
*
* @OA\Tag(
*     name="Observers",
*     description="Everything about your Observers" 
* )
*/ 
class Observers extends BaseEntity
{
    	/**
	 * @OA\Property(		 		 		 
	 *     description="observer_id",
	 *     title="observer_id",
	 *     type="string",
	 * 	   format="-",	 
	 * 	   nullable=false,
	 * 	   maxLength=6,
	 * )
	 *		 
	 */
	private $observer_id;
	/**
	 * @OA\Property(		 		 		 
	 *     description="user_id",
	 *     title="user_id",
	 *     type="string",
	 * 	   format="-",	 
	 * 	   nullable=false,
	 * 	   maxLength=6,
	 * )
	 *		 
	 */
	private $user_id;
	/**
	 * @OA\Property(		 		 		 
	 *     description="regional_unit_id",
	 *     title="regional_unit_id",
	 *     type="string",
	 * 	   format="-",	 
	 * 	   nullable=false,
	 * 	   maxLength=6,
	 * )
	 *		 
	 */
	private $regional_unit_id;
	/**
	 * @OA\Property(		 		 		 
	 *     description="observer_name",
	 *     title="observer_name",
	 *     type="string",
	 * 	   format="-",	 
	 * 	   nullable=false,
	 * 	   maxLength=50,
	 * )
	 *		 
	 */
	private $observer_name;
	/**
	 * @OA\Property(		 		 		 
	 *     description="observer_phone",
	 *     title="observer_phone",
	 *     type="string",
	 * 	   format="-",	 
	 * 	   nullable=false,
	 * 	   maxLength=20,
	 * )
	 *		 
	 */
	private $observer_phone;
	/**
	 * @OA\Property(		 		 		 
	 *     description="observer_img",
	 *     title="observer_img",
	 *     type="string",
	 * 	   format="-",	 
	 * 	   nullable=false,
	 * 	   maxLength=255,
	 * )
	 *		 
	 */
	private $observer_img;
	/**
	 * @OA\Property(		 		 		 
	 *     description="created_at",
	 *     title="created_at",
	 *     type="string",
	 * 	   format="-",	 
	 * 	   nullable=false,
	 * )
	 *		 
	 */
	private $created_at;
	/**
	 * @OA\Property(		 		 		 
	 *     description="updated_at",
	 *     title="updated_at",
	 *     type="string",
	 * 	   format="-",	 
	 * 	   nullable=false,
	 * )
	 *		 
	 */
	private $updated_at; 
}
/**
 *
 * @OA\RequestBody(
 *     request="Observers",
 *     description="Observers object that needs to be added", 
 *     @OA\JsonContent(ref="#/components/schemas/Observers"),
 *     @OA\MediaType(
 *         mediaType="application/x-www-form-urlencoded",
 *         @OA\Schema(ref="#/components/schemas/Observers")
 *     ),
 *     @OA\MediaType(
 *         mediaType="application/xml",
 *         @OA\Schema(ref="#/components/schemas/Observers")
 *     )
 * )
 */