# Pertemuan 11/Restfull API/public/assets/api.yaml
openapi: 3.0.0
paths:
...
  /industryPollutions:
    get:
      tags:
        - IndustryPollutions
      summary: 'Find list IndustryPollutions'
      description: 'Returns list of IndustryPollutions'
      operationId: getIndustryPollutions
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/IndustryPollutions' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/IndustryPollutions' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/IndustryPollutions' } }
                type: object
        '404':
          description: 'IndustryPollutions not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - IndustryPollutions
      summary: 'Add a new IndustryPollutions to the store'
      operationId: addIndustryPollutions
      requestBody:
        $ref: '#/components/requestBodies/IndustryPollutions'
      responses:
        '201':
          description: 'Created IndustryPollutions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndustryPollutions'
            application/xml:
              schema:
                $ref: '#/components/schemas/IndustryPollutions'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/industryPollutions/{id}':
    get:
      tags:
        - IndustryPollutions
      summary: 'Find IndustryPollutions by ID'
      description: 'Returns a single IndustryPollutions'
      operationId: getIndustryPollutionsById
      parameters:
        -
          name: id
          in: path
          description: 'ID of IndustryPollutions to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndustryPollutions'
            application/xml:
              schema:
                $ref: '#/components/schemas/IndustryPollutions'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'IndustryPollutions not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - IndustryPollutions
      summary: 'Update an existing IndustryPollutions'
      operationId: updateIndustryPollutions
      parameters:
        -
          name: id
          in: path
          description: 'IndustryPollutions id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/IndustryPollutions'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'IndustryPollutions not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - IndustryPollutions
      summary: 'Deletes a IndustryPollutions'
      operationId: deleteIndustryPollutions
      parameters:
        -
          name: id
          in: path
          description: 'IndustryPollutions id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /observers:
    get:
      tags:
        - Observers
      summary: 'Find list Observers'
      description: 'Returns list of Observers'
      operationId: getObservers
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Observers' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Observers' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/Observers' } }
                type: object
        '404':
          description: 'Observers not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - Observers
      summary: 'Add a new Observers to the store'
      operationId: addObservers
      requestBody:
        $ref: '#/components/requestBodies/Observers'
      responses:
        '201':
          description: 'Created Observers'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Observers'
            application/xml:
              schema:
                $ref: '#/components/schemas/Observers'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/observers/{id}':
    get:
      tags:
        - Observers
      summary: 'Find Observers by ID'
      description: 'Returns a single Observers'
      operationId: getObserversById
      parameters:
        -
          name: id
          in: path
          description: 'ID of Observers to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Observers'
            application/xml:
              schema:
                $ref: '#/components/schemas/Observers'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'Observers not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - Observers
      summary: 'Update an existing Observers'
      operationId: updateObservers
      parameters:
        -
          name: id
          in: path
          description: 'Observers id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/Observers'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Observers not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - Observers
      summary: 'Deletes a Observers'
      operationId: deleteObservers
      parameters:
        -
          name: id
          in: path
          description: 'Observers id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
...
    IndustryPollutions:
      title: IndustryPollutions
      description: IndustryPollutions
      properties:
        industry_pollution_id:
          nullable: false
          title: industry_pollution_id
          description: industry_pollution_id
          type: string
          format: '-'
          maxLength: 6
        observer_id:
          nullable: false
          title: observer_id
          description: observer_id
          type: string
          format: '-'
          maxLength: 6
        industry_pollution_title:
          nullable: false
          title: industry_pollution_title
          description: industry_pollution_title
          type: string
          format: '-'
          maxLength: 70
        industry_pollution_description:
          nullable: false
          title: industry_pollution_description
          description: industry_pollution_description
          type: string
          format: '-'
        industry_pollution_longitude:
          nullable: false
          title: industry_pollution_longitude
          description: industry_pollution_longitude
          type: string
          format: '-'
          maxLength: 100
        industry_pollution_lattitude:
          nullable: false
          title: industry_pollution_lattitude
          description: industry_pollution_lattitude
          type: string
          format: '-'
          maxLength: 100
        industry_pollution_type:
          nullable: false
          title: industry_pollution_type
          description: industry_pollution_type
          type: string
          format: '-'
        industry_pollution_img:
          nullable: false
          title: industry_pollution_img
          description: industry_pollution_img
          type: string
          format: '-'
          maxLength: 255
        industry_pollution_status:
          nullable: false
          title: industry_pollution_status
          description: industry_pollution_status
          type: string
          format: '-'
        industry_pollution_action:
          nullable: false
          title: industry_pollution_action
          description: industry_pollution_action
          type: string
          format: '-'
        industry_pollution_action_img:
          nullable: false
          title: industry_pollution_action_img
          description: industry_pollution_action_img
          type: string
          format: '-'
          maxLength: 255
        created_at:
          nullable: false
          title: created_at
          description: created_at
          type: string
          format: '-'
        updated_at:
          nullable: false
          title: updated_at
          description: updated_at
          type: string
          format: '-'
      type: object
    Observers:
      title: Observers
      description: Observers
      properties:
        observer_id:
          nullable: false
          title: observer_id
          description: observer_id
          type: string
          format: '-'
          maxLength: 6
        user_id:
          nullable: false
          title: user_id
          description: user_id
          type: string
          format: '-'
          maxLength: 6
        regional_unit_id:
          nullable: false
          title: regional_unit_id
          description: regional_unit_id
          type: string
          format: '-'
          maxLength: 6
        observer_name:
          nullable: false
          title: observer_name
          description: observer_name
          type: string
          format: '-'
          maxLength: 50
        observer_phone:
          nullable: false
          title: observer_phone
          description: observer_phone
          type: string
          format: '-'
          maxLength: 20
        observer_img:
          nullable: false
          title: observer_img
          description: observer_img
          type: string
          format: '-'
          maxLength: 255
        created_at:
          nullable: false
          title: created_at
          description: created_at
          type: string
          format: '-'
        updated_at:
          nullable: false
          title: updated_at
          description: updated_at
          type: string
          format: '-'
      type: object
...
  requestBodies:
...
    IndustryPollutions:
      description: 'IndustryPollutions object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/IndustryPollutions'
        application/xml:
          schema:
            $ref: '#/components/schemas/IndustryPollutions'
        application/json:
          schema:
            $ref: '#/components/schemas/IndustryPollutions'
    Observers:
      description: 'Observers object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Observers'
        application/xml:
          schema:
            $ref: '#/components/schemas/Observers'
        application/json:
          schema:
            $ref: '#/components/schemas/Observers'
...